+++ Rubrik/Public/Get-RubrikVMwareDatacenter.ps1
@@ -14,7 +14,7 @@ function Get-RubrikVMwareDatacenter
       GitHub: mwpreston
             
       .LINK
-      https://rubrik.gitbook.io/rubrik-sdk-for-powershell/command-documentation/reference/get-rubrikVMwareDatacenter
+      https://rubrik.gitbook.io/rubrik-sdk-for-powershell/command-documentation/reference/get-rubrikvmwaredatacenter
             
       .EXAMPLE
       Get-RubrikVMwareDatacenter