+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+* Modified Link for `Protect-RubrikVolumeGroup` to lowercase
+* Modified `Remove-RubrikVMSnapshot` to support -Confirm before performing action
+* Updated help for all fileset and fileset template related cmdlets as per [Issue 284](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/284)
 * Changed example documentation for `New-RubrikDatabaseMount`
 * Added `-DetailedObject` parameter to `Get-RubrikLogShipping`
 * Updated unit tests for `Get-RubrikLogShipping` to include test for `-DetailedObject` parameter
@@ -28,6 +31,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+* Added new unit tests: `CommentBasedHelp`, `ObjectDefinitions`, `Connect-Rubrik`, `Disconnect-Rubrik`
+* Added `Remove-RubrikFilesetSnapshot`, `RemoveRubrikDatabaseSnapshots`, `Remove-RubrikHyperVSnapshot`, `Remove-RubrikManagedVolumeSnapshot`, `Remove-RubrikNutanixVMSnapshot`, and `Remove-RubrikVolumeGroupSnapshot` along with associated unit tests as outlined in [Issue 309](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/309) *NOTE Remove-RubrikDatabaseSnapshots deletes ALL snapshots for a given database - there is currently no endpoint to support the deletion of a single snapshot*
 * Added `Suspend-RubrikSLA` and `Resume-RubrikSLA`
 * Added `Get-RubrikEventSeries` to now parse the event_series API rather than events as per [Issue 580](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/580)
 * Added unit tests to cover `Get-RubrikEventSeries` and changes to `Get-RubrikEvent`